ARCO-149: refactor migrations and postgres methods
