+++ internal/metamorph/server.go
@@ -173,14 +173,13 @@ func (s *Server) PutTransaction(ctx context.Context, req *metamorph_api.Transact
 	statusReceived := metamorph_api.Status_RECEIVED
 
 	// Convert gRPC req to store.StoreData struct...
-	callback := store.StoreCallback{
-		CallbackURL:   req.GetCallbackUrl(),
-		CallbackToken: req.GetCallbackToken(),
-	}
 	sReq := &store.StoreData{
-		Hash:              hash,
-		Status:            statusReceived,
-		Callbacks:         []store.StoreCallback{callback},
+		Hash:   hash,
+		Status: statusReceived,
+		Callbacks: []store.StoreCallback{{
+			CallbackURL:   req.GetCallbackUrl(),
+			CallbackToken: req.GetCallbackToken(),
+		}},
 		FullStatusUpdates: req.GetFullStatusUpdates(),
 		RawTx:             req.GetRawTx(),
 	}
@@ -210,14 +209,13 @@ func (s *Server) PutTransactions(ctx context.Context, req *metamorph_api.Transac
 		timeout = txReq.GetMaxTimeout()
 
 		// Convert gRPC req to store.StoreData struct...
-		callback := store.StoreCallback{
-			CallbackURL:   txReq.GetCallbackUrl(),
-			CallbackToken: txReq.GetCallbackToken(),
-		}
 		sReq := &store.StoreData{
-			Hash:              hash,
-			Status:            statusReceived,
-			Callbacks:         []store.StoreCallback{callback},
+			Hash:   hash,
+			Status: statusReceived,
+			Callbacks: []store.StoreCallback{{
+				CallbackURL:   txReq.GetCallbackUrl(),
+				CallbackToken: txReq.GetCallbackToken(),
+			}},
 			FullStatusUpdates: txReq.GetFullStatusUpdates(),
 			RawTx:             txReq.GetRawTx(),
 		}
